requirements.txt
appdirs==1.4.3
astroid==1.6.5
backports.functools-lru-cache==1.5
backports.os==0.1.1
configparser==3.5.0
enum34==1.1.6
fs==2.1.2
future==0.17.1
futures==3.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.3
python-slugify==1.2.6
pytz==2018.7
singledispatch==3.4.0.3
six==1.11.0
termcolor==1.1.0
typing==3.6.6
Unidecode==1.0.23
wrapt==1.10.11
